pyproject.toml
[project]
name = "data-charts-api"
version = "0.1.0"
description = "api for data charts project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.0.3",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.3",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.3",
    "requests>=2.32.3",
    "ruff>=0.7.2",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
